# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: MCT Client Portal API
  description: Maine CyberTech Client Portal API. Bearer token authentication required unless noted as public.
  version: 1.0.0
  contact:
    name: MCT Support
    email: support@mainecybertech.com
servers:
  - url: https://api.mainecybertech.com
    description: Production
  - url: https://api.mainecybertech.us
    description: Dev
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api-keys:
    get:
      tags:
        - Admin
      summary: List API keys
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
    post:
      tags:
        - Admin
      summary: Create API key
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
  /api-keys/{id}:
    patch:
      tags:
        - Admin
      summary: Update API key
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
    delete:
      tags:
        - Admin
      summary: Delete API key
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
...
  /store/categories:
    get:
      tags:
        - Store
      summary: List categories (public)
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security: []
  /store/categories/{slug}:
    get:
      tags:
        - Store
      summary: Get category with products (public)
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
  /store/products:
    get:
      tags:
        - Store
      summary: List products (public)
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security: []
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
  /store/products/{slug}:
    get:
      tags:
        - Store
      summary: Get product by slug (public)
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
  /store/promotions:
    post:
      tags:
        - Store
      summary: Create promotion
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
    get:
      tags:
        - Store
      summary: List promotions
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
  /store/promotions/active:
    get:
      tags:
        - Store
      summary: List active promotions (public)
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security: []
  /store/promotions/admin:
    get:
      tags:
        - Store
      summary: List promotions (admin)
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
  /store/promotions/{id}:
    patch:
      tags:
        - Store
      summary: Update promotion
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
    delete:
      tags:
        - Store
      summary: Delete promotion
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
  /store/quotes:
    post:
      tags:
        - Store
      summary: Submit quote request
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security: []
    get:
      tags:
        - Store
      summary: List quote requests (admin)
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
...